% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcsqlite-package.R
\name{adbcsqlite}
\alias{adbcsqlite}
\title{ADBC SQLite3 Driver}
\usage{
adbcsqlite()
}
\value{
An \code{\link[adbcdrivermanager:adbc_driver_void]{adbcdrivermanager::adbc_driver()}}
}
\description{
ADBC SQLite3 Driver
}
\examples{
adbcsqlite()

}
